replay viewer: add 1/2x playback speed; Space pause verified on every shipped page (#1)
The replay transport gains a 0.5x speed: command '5' selects a replay-only
half speed (ReplayHalfSpeedIndex) that advances one frame every other
presentation frame, '-' floors there, the chrome grows a 0.5x chip, and the
frame JSON's sp field carries 0.5 while it is active (buildStateJson /
buildViewerPacket speed is a float now). The speed vocabulary lives in
src/factory_commons/replays.nim so the native tests can exercise it; the
wasm playhead in replay-viewer/factory_commons_replay.nim consumes it and
toggles halfPhase once per frame. wire_constants.nim rides 0.5 ahead of the
engine's integer PlaybackSpeeds (the CTF_WIRE global keeps its Dockerfile-
pinned name), and chrome_common's fallback SPEEDS and speed-to-command map
gain the 0.5 entry. The only shipped page (index.html, built from
client/replay_broadcast.html) already binds Space to play/pause directly —
this repo ships no shell page, so no command-channel forwarding was needed.

+++ src/factory_commons/broadcast.nim
@@ -127,7 +127,7 @@ proc buildStateJson*(
   model: HudModel,
   events: JsonNode,
   playing: bool,
-  speed: int,
+  speed: float,
   looping: bool,
   transportEnabled: bool,
   leadSeries: seq[array[3, int]] = @[],
